[sestarnix] Include namespace path in fs_node_permission audits
Allow callers of fs_node_permission() to supply an additional
AuditContext<> to be include if the access check triggers any audit
logs to be generated.

This is used within FsNode::open() to provide include the path of
the NamespaceNode via which the node is being opened, to aid
diagnosis of denials.

Bug: 380855359
Change-Id: I9396bc2391fdbffd950613d265f3412cd1fc08d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374882
Fuchsia-Auto-Submit: Wez <wez@google.com>
Reviewed-by: Ambre Williams <ambre@google.com>
Commit-Queue: Wez <wez@google.com>