[wlan][policy] Fix metric logging on failed roam
Investigation doc: go/haydennix:missing-failed-roam-metrics

While monitoring the WLAN roaming metrics dashboard, I noticed an
inconsistency that led me to believe the detailed roam result metric was
not being logged correctly for a failed roam. The cause was that, following a failed roam
result, the wlancfg client state machine logs a disconnect event to the
telemetry module before logging the roam result. This was to ensure the
internal state of both the client state machine and the telemetry module
contained up-to-date information for before any logging.

In the telemetry module, there was a conditional to prevent the
roam result metrics from being logged if the internal telemetry state
was not "Connected". This was intended to prevent logging roam results
that come in after a disconnect for an unrelated reason.

In practice, a failed roam result would send a disconnect event,
changing the telemetry state to "Disconnected", and then the conditional
would prevent to the roam result metrics from being logged.

This change re-orders the conditional in the telemetry module so that we
only match the telemetry state on successful roams, because getting a
successful roam result when internal state is "Disconnected" would be
unexpected. But failed roam results are still logged, even though they come in
after a disconnect event.

Bug: None
Test: fx test wlancfg-tests, ran manual tests to verify metrics are
logging.
Multiply: wlancfg-tests
Change-Id: I5b3a1846d7db969151609f14646068a0ea296f89

Change-Id: Ib4e943c478cc59ab5e445d47d76d46151e74ba11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374302
Commit-Queue: Hayden Nix <haydennix@google.com>
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>
Reviewed-by: Naomi Mccracken <nmccracken@google.com>