[fdf][rust] Defer channel close while await is registered
With async read cancelation happening asynchronously now on
both unsynchronized dispatchers and fuchsia-async hack
dispatchers, it becomes a lot more likely that the
channel itself is dropped while the callback is still pending.

In order to correctly deal with that, the channel itself
needs to be aware that there is a pending wait so that it
can defer closing until the callback is called. Likewise,
the callback will sometimes need to close the channel.

In the process we've also made it so that the Arc allocated
for the future is saved and used in subsequent read waits,
which has the double benefit of making sure that the state
for the cancel is correct across subsequent awaits and
reduces memory allocations across the life of a channel.

Multiply: fdf_rust
Bug: 439021012
Change-Id: I29397cec60c632cf4b95d5a70b8bb6c13760eaac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1372134
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>