[shutdown-shim] Add Shutdown to Admin protocol
The RebootWatcher protocol is used to notify components about why a
reboot is about to occur. Feedback uses the included reason to help
determine its canonical "reboot reason," which is used in several
places.

However, Feedback is missing important context for shutdowns that aren't
just a reboot into product mode. For example, the device may power off
rather than reboot when high temperature is detected.

This CL is the first of many that will eventually allow Feedback to be
notified of the reason for other types of shutdowns. In particular,
this CL adds a "Shutdown" method that will allow clients to specify
both the shutdown reason and shutdown action, such as "reboot" or
"poweroff."

Clients will be migrated to use Shutdown instead of other methods which
may not allow the client to specify a reason.

For now, only reasons attached to a "reboot" action will be forwarded to
RebootWatcher clients. Future CLs will add a ShutdownWatcher protocol.

Bug: 414413282
Test: `fx test //src/power/shutdown-shim`
Test: `fx reboot` --> Feedback reboot reason "developer request"
[g-review]

Change-Id: Ic83cb095c2b534da800dd6429358434028d49833
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370968
Commit-Queue: Thomas Gales <tgales@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
API-Review: Francois Rousseau <frousseau@google.com>
Reviewed-by: Hai Bi <bihai@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>