[fxfs] Fix a compaction/delete volume race
The code was capturing a copy of the allocator info at the wrong time.

This also fixes another issue: when we're compacting we take a snapshot
of the volumes to be compacted, but they can be deleted before we
actually get to compacting them. This would have triggered a compaction
error, which is considered fatal.

Multiply: fxfs-test
Fixed: 445533482
Change-Id: Ifa8a15427114a7dde8c895104c35efb576c19c50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374583
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>