[ffx][trace] Adding check for --background trace start
This adds a check to validate that `--background` is not used
at the same time as `--output`.  Since the background trace no
longer interacts with the output file, it does not make sense,
and there is not an easy way to remember the file name until
`ffx trace stop` is called.

The correct way to use `--output` is with foreground tracing, or
use the `ffx trace stop --output`.

Bug: 446684592
Change-Id: Ic4a3fd0b22b3f90e82d311ffd087f010c4fa86b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374245
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>