[sestarnix] Fix missing-xattr fallback for fs_use_xattr root nodes
SEStarnix has a workaround to set the xattr on the root node of
`fs_use_xattr` filesystems, if it is not present, based on the
filesystem's "rootcontext=" mount option.

Previously this would fall-back to setting the filesystem's label
in the xattr for the root node, which appears to be incorrect. The
root node now falls-back to receiving the "default" SID for the
filesystem (i.e. from "defcontext=" or the "file" initial SID) and
the value is no longer persisted to the node in this case.

Bug: 445734070
Change-Id: I1e615fcdcecdb3f3ef3ad4c463c75bb4710646d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374902
Commit-Queue: Wez <wez@google.com>
Reviewed-by: Ambre Williams <ambre@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>