[starnix] Add Augmented type for io_uring ring buffers
This change introduces an `Augmented<T, A>` type that allows a primary
value of type `T` to be augmented with an optional auxiliary value of
type `A`.

This is used in `recvmsg` to allow the `iovec` from the `msghdr` to be
overridden by a `UserBuffer`. This is necessary to support `io_uring`
with ring buffers, where the buffer is provided by the ring buffer
itself, and not by the `msghdr`.

The `recvmsg` syscall is split into `sys_recvmsg` and `recvmsg_impl` to
allow `io_uring` to call into the implementation with an augmented
`msghdr`.

Bug: 297431387
Change-Id: Iea349ae59b008bfcc65ce71a2a088b0f5e8266b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374021
Commit-Queue: Benjamin Lerman <qsr@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Nathaniel Manista <nathaniel@google.com>