[owners] Remove jaeheon@google.com from src/chromium/web_runner_tests/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the
OWNERS file tidy. We ask that you do not ignore this change and either
approve it if the OWNER is no longer active or deny it if the OWNER is
continuing to review CLs (including under a different affiliation).
This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the
CL) and the bot will abandon it.

See the owner's recent review activity for context:
https://fuchsia-review.googlesource.com/q/jaeheon@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I04d2f5fc7d0df490eff96e9fefd381a4c340218b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374991
Fuchsia-Auto-Submit: Owners-Cleanup-Prod <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Emircan Uysaler <emircan@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>