[unwinder][cfi] Size memory reads based on current ELFCLASS
The CFI Parser previously assumed all reads would be for 64 bit values,
which is not the case when parsing CFI from a 32 bit restricted mode
process which may contain both CFI and EHABI unwinding instructions.

In particular we need to be sure when reading values from the stack or
from memory that we're only reading the expected 4 byte values when
we're dealing with 32 bit code.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a69640b78d085ba6762e473e9e5f16e4e2ded
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371886
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>