[kernel][bti] Flag critical process teardown
We need a reliable way to determine if the system is on its way down due
to a critical process or job being killed so we can avoid emitting BTI
OOPS messages when driver_manager is taken down. Unfortunately, the
previous solution of using GetCriticalProcessKoid() was unreliable
because the value was set at the end of the process tree being released.
In certain circumstances, the drivers holding BTIs would be cleaned up
before it would be set. This change marks the very beginning of such a
chain so that we can reliably check if that state is true.

Fixed: 121831
Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs

Change-Id: I0424fcba2fd82bd15985e0fba6d109f559cdfcce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/813821
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Christopher Anderson <cja@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>