[unwinder] Handle 32 bit register transition better
This moves the transition handling to the start of processing of the
first 32 bit frame, rather than at the end of the last 64 bit frame.
This way, we can more easily examine the ELF module for the current PC
value, which can correctly tell us whether or not it's 32 bit or 64 bit.

Fixed: 441470079
Bug: 416040479
Bug: 395606674
Change-Id: I6a6a6964a8b94215e1454a9e86ecd9338d5e3568
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1375421
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>