[kernel][phys] Replace ZirconInfo ELF note with e_entry pointer
This switches to using the ELF header's e_entry field to address
kZirconAbiSpec rather than getting its contents into the payload
of an ELF note, with the real entry point stored in ZirconAbiSpec
as a (relocated) function pointer.  Avoiding ELF notes means no
extra constraints on the type, such as on its alignment.  It can
be any `constinit const` object, even RELRO, as it now is to have
a relocated function pointer for the entry point.

It would also be easy to do a symbol lookup of an exported
symbol, as is done for `__cfi_check` in phys modules.  But the
metadata for symbols and the lookup work is more overhead than a
simple pointer plumbed via `-e` as the ELF entry point, as is the
ELF note metadata.  Even several relocated pointers inside
kZirconAbiSpec still wind up cheaper than the other metadata.

This gets rid of all the pseudo-generic ZirconInfo note machinery
in ElfImage, replacing it all with just having physboot fetch a
pointer into the image based on the e_entry value.

Cq-Include-Trybots: luci.turquoise.global.try:run-postsubmit-tryjobs
Change-Id: I0c1dbc3fb010cd1f262e2f9fda7a7d32e836399c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1373061
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>