[windowed-stats] Refactor `Statistic` traits.
This change consolidates the `Sampler` and `Fill` traits into the
`Statistic` trait. This means some items have been consolidated
elsewhere too, like the `fold` function for `MatrixSampler`.

This change also introduces the `SerialStatistic` trait. Though this is
an addition, it greatly simplifies trait bounds and implicitly relates
the associated types of a `Statistic` with the `BufferStrategy`
implementation of its `Semantic`. Before this change, `Statistic` types
had a blanket implementation of `BufferStrategy`, but this required more
verbose and redundant bounds.

Test: unit tests
Change-Id: I2d585e3afd1d0a2c72c1f040b7af190f84711da5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371383
Reviewed-by: Kiet Tran <kiettran@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>