[kernel] Remove k fibo kernel command
Removes the fibonacci kernel command. Nick and Carlos requested this
cleanup. This kernel command is not used in CI/CQ nor is it run by
developers directly.

Instead of integrating fibo into the thread unittests, this change
removes it entirely.

k fibo originally spawned threads recursively that each returned the nth
fibonacci number when called with n as an argument. It was intended to
test the performance and reliability of creating many threads in an
interactive context.

Bug: 447206619
Change-Id: I5800854c99bdacff72c177cdbb5bb71ede15bbcf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1376392
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Ethan Williams <ethanws@google.com>
Fuchsia-Auto-Submit: Ethan Williams <ethanws@google.com>