[sestarnix] Add audit netlink userspace tests
Added some basic userspace tests that pass on Linux and should be fixed
on Starnix as well. Other tests that use a SELinux policy will be added.

The tests cover some edge cases that were identified while working on
the Starnix audit logger.

Added `audit=1` to the `userspace_test_container` kernel command line
to enable the backlog in the audit logger. This is also set for running
on Linux.

One test is skipped because of a really rare hang when doing a receive.

Bug: 443684136
Change-Id: Id6dbeda8121a63cbf346104c1f3a52625a926fd6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1369537
Commit-Queue: Rareş Constantin <rarescon@google.com>
Reviewed-by: Wez <wez@google.com>