[sestarnix] Add netlink audit policy and tests
Added the netlink audit policy which is used for the netlink audit
userspace tests. The tests include the socket create cases and a send
test suite which verifies different security labels for the netlink
audit socket class.

All tests pass on Linux, two of the send tests are marked to expect
failure on Starnix.

Bug: 443684136
Change-Id: I0cebbf5d809f8fcda2d33ab7c89104dc0a1775e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370976
Commit-Queue: Rareş Constantin <rarescon@google.com>
Reviewed-by: Wez <wez@google.com>