[sestarnix] Audit log userspace testing
Added an audit log checker, a test listener that is constructed with an
expectations file which is parsed before running any tests. If a test is
present in the expectations file, the audit logs produced by it will be
checked. The audit logs of a specific test are guarded by a start
sentinel and an end sentinel messages. If there is a message printed
before the start sentinel, it means that a previous test or setup
produced a policy violation that must be checked.

Each test may add its own audit log expectations as audit strings inside
a JSON array in the `audit_expectations.json` file. The expectations
should be generated using the `run_on_linux.py` script to extract what
audit logs are printed on Linux. The audit checker can be initialized
with the debug flag set to true to only print any audit messages in
between the sentinels and skip checks.

Bug: 443684136
Change-Id: If47c8fc218b6e6ac2f4b8258c3d23607e2debd86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1364172
Commit-Queue: Rareş Constantin <rarescon@google.com>
Reviewed-by: Wez <wez@google.com>