[fxfs] Fix incorrect last object ID after volume installation
When a new volume is installed, we create a new object to keep ownership
of the extents we want to delete post-install. Although we set the
object count correctly, we were not setting the last object ID in the
persisted store info. The in-memory state of the object store was also
not being updated. This was causing issues when creating new objects in
the installed volume, since transactions would attempt to insert an
object that already existed.

To ensure this does not regress, we add a new check in fsck to ensure
that the next object ID an ObjectStore will use is not already assigned
to an object.

Bug: 441562849
Test: fx test fxfs-tests --test-filter="*install*"

Change-Id: I892bfde443efe46317a6d33c577195f7eb72e6c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1376323
Commit-Queue: Brandon Castellano <bcastell@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>