[fxtest] Ensure package lists are updated for device tests.
Without this change, it is possible to add a package to the build that
is not reflected in the package lists that will be published to the TUF
repository for the device, at least in cases where the assembly
process uses Bazel.

With this change we refresh the package lists files to ensure any new
additions are caught and published correctly for testing purposes. We
can do this without needing to build the entire updates package like we
need to do for e2e tests.

Fixed: 442839521
Change-Id: I0213786ea7b2ce161ade26e2bf2cd197c926128e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1376821
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>