[runtime][doc] Improve the documentation of `fuchsia_runtime::utc_time`.
The documentation of the function did not emphasize the subtleties of
the use of UTC timestamps, some of which are unique to Fuchsia.

This leads to frequent misuse of the function in contexts where it is
too easy to transplant assumptions about UTC from other systems.

Bug: 447454745
Change-Id: I6aacf07442b8caca89fe32b252147dc3a084995b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1377489
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>