[elflib] Parse and upcast 32 bit Symtabs
This follows the same procedure as dealing with 32 bit ELF headers,
where we handle the 32 bit cases and then upcast the Elf32_Sym objects
to Elf64_Sym objects so callers don't have to think about it.

Adding tests for the Dynamic symbol table is pretty tricky because we'd
have to somehow construct a DT_GNU_HASH entry which is a bloom filter
binary tree, just to calculate the size of the symbol table. Until we
have something upstreamed to LLVM that gets us the count of symbols I
don't think it's going to be very easy to make that happen.

Bug: 395606674
Change-Id: I6a6a69646a2c009c51a5ca99160661af803aedc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1375423
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>