[sdk][driver] Initiate driver handle transfer before connect
This changes the C++ namespace bindings to call `fdf_token_transfer`
before attempting to connect, ensuring that the server will not have
to wait for the token to be available when it receives the
connection request. This also opens the door to making this process
synchronous in the future.

Change-Id: I545ac97dc973e903de148039a5b4b5d6410e0406
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1377322
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Megan Batty <mgnb@google.com>