[kernel][thread] Enable spinlock_test in k ut intensive_thread
The spinlock_test was disabled when the `k thread` command was migrated
to the kernel unittests. It was disabled because it was originally
failing.

This CL fixes the failure by adding an AutoPreemptDisabler. It
then re-enables the test.

Bug: 439836795
Change-Id: Ib0e577dd10a31f10b5fdf3a063df816a203979f2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1376688
Fuchsia-Auto-Submit: Ethan Williams <ethanws@google.com>
Commit-Queue: Ethan Williams <ethanws@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>