[storage] Allow empty attributes in macro
The vfs::attributes! macro enables ergonomic and efficient construction
of NodeAttributes2 at compile time:

    attributes!(
        requested_attributes,
        Mutable { creation_time: 123, modification_time: 456 },
        Immutable { content_size: 789 }
    )

However, if either the Mutable or Immutable blocks are empty, the macro
expansion is syntactically invalid due to the presence of an extra
comma.

This change corrects the expansion, allowing for either block to be
empty, making the construction of attributes that only include Mutable
attributes possible without the introduction of a specialized
mutable_attributes! macro.

    attributes!(
        requested_attributes,
        Mutable { creation_time: 123 },
        Immutable {}
    )

Test: builds
Change-Id: Icf864188b00b02567b8e6e3533f86028efe6b4c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1372504
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Jeremiah Griffin <jzgriffin@google.com>