[storage][ext4] Add ext4 directory entry types
This change adds the following:
- ExtDirectory: A new Directory type, based on Simple, which represents
  an immutable ext4 filesystem directory.
- ExtFile: A new File type, based on VmoFile, which represents an
  immutable ext4 filesystem file.

The purpose of this change is to allow ext4 attributes to be surfaced in
the mapped fuchsia.io nodes from the underlying ext4 inodes. In
particular, the access mode, owner UID, and owner GID attributes are
surfaced in this way.

Bug: 439904101
Test: storage-ext4-parser-tests
Change-Id: Iffbc516656f340079138300ad63c1144bbfee1c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1372507
Commit-Queue: Jeremiah Griffin <jzgriffin@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>