[kernel][vm] Add delayed reuse to PMM
This commit introduces a minor PMM feature, delayed reuse.  See
https://fxbug.dev/446957856 and https://fxbug.dev/445524948.

The idea behind this feature is that when freeing pages to the PMM, we
can either make them available for immediate reuse or delay their
reuse until after all other previously freed pages have been used.

Generally speaking, immediate reuse has performance benefits because a
recently freed page is more likely to be present in some sort of cache
(or TLB).  Because of the performance benefits, immediate reuse is the
default behavior in most build configurations.

However, delayed reuse can be preferable in some specific situations.

Consider the case of pinned memory used for DMA.  If a software bug in
a driver results in prematurely unpinning the memory (i.e. before the
DMA engine has been properly shut down), we can end up in a
use-after-free scenario.  If the freed page is immediately reused,
even before the errant DMA occurs, we could, in the case of it being a
source page, incorrectly expose the new page's content (think DMA to
network or disk), or, in the case of a destination page, stomp/corrupt
its new contents.  Either way, it's a bad time.

By delaying the reuse of pages used in DMA transfers, we increase the
probability that the page is on the free list when a bad DMA occurs.
When a page is corrupted while on the free list, the impact is
mitigated and PMM checker is more likely to detect the issue.

This change merely adds the feature.  It will be enabled in future
changes.

Bug: 445524948
Bug: 446957856
Change-Id: I3064924f764aa2503a8c54f665d4c91d72302fc5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1374087
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>