[netlink] Mover worker initialization to the worker thread
Previously netlink worker was initialized on a thread different from
the thread where the worker runs, which means that all worker internals
had to be `Send`. This CL moves worker initialization to the worker
thread, which allows to avoid `Send` and `'static` for parameters passed
to the worker. In the future this will allow to add dependency on
`&CurrentTask` in the worker

Bug: 441781721
Change-Id: I5ecff5d27906b79fbe48b3236f0716185e6e5600
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1377486
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>