[starnix][io_uring] Implement recvmsg with provided buffers
This change implements support for `IORING_OP_RECVMSG` with the
`IOSQE_BUFFER_SELECT` flag in io_uring. This allows clients to use a
registered ring of buffers for receiving data, which can be more
efficient as it avoids extra data copies.

The implementation adds a `IoUringProviderRingBuffer` to manage the
registered buffers and a `claim_next_buffer` method to get the next
available buffer from the ring. The `recvmsg` operation is updated to
use this buffer when `IOSQE_BUFFER_SELECT` is set.

To support this, `recvmsg_impl` in `vfs/socket/syscalls.rs` was
refactored to take a `MsgHdrRef`, which can represent either a user-space
pointer to a `msghdr` or an in-kernel `msghdr` with an alternate buffer.

Bug: 297431387
Depends-on: turquoise-internal:I854a7dad0641783147aed5e1241f3c4dcaf5ae91
Change-Id: Iacede38d82ccb1adcc1fabf19d906cc0c15a3383
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1377881
Dependencies-Satisfied: CL Deps Checker <cl-deps-checker@fuchsia-internal-service-accts.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>