[memory_monitor2] Reduce thread synchronization overhead
Several functions in memory_monitor2 consume Send arguments while
never needing to do any cross-thread work. This was a convenient
default stance, because that means we can fairly easily spawn tasks
from anywhere in the call tree, but we don't use this ability, the
code is not too idiomatic because of this, and we might pessimize the
code in some cases due to the unneeded synchronization mechanisms.

This change simplifies the situation a bit by removing unused/unlikely
to be used synchronization code.

Change-Id: I0d37684a3f533fdb497626cdef14673d0b137b70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1373923
Commit-Queue: Guillaume Chelfi <chelfi@google.com>
Reviewed-by: Étienne J. Membrives <etiennej@google.com>