[assembly] allow duplicate firmware images
Some boards have slotted firmware so want to specify both an A and B
firmware image, but the OTA package creation only wants one copy (since
the slot is applied at runtime based on A/B state) and providing both
was causing build failures.

This CL allows firmware images to automatically de-duplicate if the OTA
type and image contents are identical. The existing "overwrite_files"
setting didn't quite work here because we don't want to unconditionally
override, but only if the contents are exactly the same.

Bug: b/434935097
Change-Id: I65f7903acc9af0c388656bd6e3dac31ae81a81a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1377958
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: David Pursell <dpursell@google.com>