[f2fs_reader] Add verity files to test image.
Adds an inline verity file, and large verity file to get a merkle tree
with multiple layers, a more normally sized file and one that is in the
encrypted directory.

Also adds some reverse lookups after removing the crypt key so that
updating the encrypted file names in the tests is easier. Outputs are
formatted to copy and paste them over to the test directly.

Bug: 399727919
Change-Id: I35f916a3ab8a192b18865619deb3c4a78d47611d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1377241
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>