[ffx][trace] Remove using target specifier as output name
There was some older code that used the target specifier as the
default output file name if one was not specified.

In a recent change the default was removed from the Argh declaration
in order to validate that the output and background options could not
both be specified.

This removes to using the target spec.

Change-Id: Id5d008d977b88797433022bea7820c13e8f1a88e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1378522
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jia Li <lizhijia@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>