[zircon] Simplified barrier API in @next
This adds a simple barrier API that asks the kernel to issue a memory
and/or instruction stream barrier on all running cores in the system.
This version of the API does not include any facilities for specifying
the scope of the barrier, all barriers are issued on all running cores.

Bug: 297526152
Change-Id: Ib721521f9d37d1966c4d62dc58171a8c8b1fca1c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1361558
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
API-Review: Rasha Eqbal <rashaeqbal@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>