[unwinder] Use register values recovered by CFI in EHABI
Some binaries can ship with both CFI (typically in the form of .eh_frame
sections) and EHABI which can cover the same addresses. When this is the
case, we want to process the CFI first and recover all that we can, and
if the CFI was unable to recover the PC or return address then we can
continue on with the recovered values with the EHABI unwinder.

Bug: 395606674
Bug: 416040479
Change-Id: I6a6a69644acb20f73eb3477364b7ff1cb909fb6f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371889
Reviewed-by: James Robinson <jamesr@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>