[kernel][arm64][armv9] Hard set the CNTHCTL_EL2 register when dropping from EL2
Instead of setting and clearing bits we need in this register, hard set
it to what we need. This removes any implicit configuration a core may
have coming out of the bootloader or firmware.

This implicitly clears some of the newer bits that are enabled by
FEAT_ECV which may cause some additional traps to EL2 when accessing
virtual timer registers.

Change-Id: I536457457b206c126bc4f03a6a37c9b2a1ccc9da
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1380061
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>