[assembly] Simplify process of making test PBs
We use new templates/rules to simplify the process of making PBs for
testing. This reduces the amount of boilerplate and makes it harder
to make mistakes.

We also add an option for a product_bundle_test_group to directly
depend on a product bundle, so the user does not have to manually
add it to product_bundle_labels or build it explicitly.

Gemini can bootstrap a new product bundle with:
  Follow the instructions at @examples/assembly/test_product_bundle/defining_test_product_bundle.md
  to create a new test product bundle in bundles/buildbot/virtualization/BUILD.gn

Bug: 447431754
Change-Id: I0f6c76e952357d07cd4260dbfa2f50d1d6f8fb01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1377981
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Knobloch <aknobloch@google.com>