[fxtest] Reland: Ensure package lists are updated for device tests.
This is a reland of a previously submitted CL that added support for
rebuilding the package lists when running device tests. It turns out
that on many products rebuilding just the package lists is still an
expensive operation, when what we really want is to build the minimal
set of packages that is necessary to run the selected set of tests.

This change now checks if it is necessary to rebuild the package lists,
which happens when a requested test package does not have a merkle hash
listed in the package targets in the package repository. If the package
is listed, we just rebuild the package, otherwise we do rebuild the
package lists.

Bug: 442839521
Fixed: 448165046
Change-Id: I89573a15fee3e73481b2154646333e65f040afe2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1380581
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>