[shutdown-shim][fidl] Add ShutdownWatcherRegister
Compared to RebootWatcher, ShutdownWatcher allows clients to be notified
of both a shutdown reason as well as a shutdown action. In particular,
clients can now know why the device is powering off, rebooting to
bootloader or rebooting to recovery.

In order for ShutdownWatchers to benefit from this change, shutdown
requesters must be migrated to call Shutdown instead of the deprecated
methods. This migration will be done in future CLs.

Bug: 414413282
Test: `fx test //src/power/shutdown-shim`
Test: `fx reboot` --> Feedback reboot reason "developer request"

Change-Id: I4b5ad679a049611030a641a447c2d9e845cc53c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1371541
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Thomas Gales <tgales@google.com>
API-Review: Francois Rousseau <frousseau@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>