[fidl_codec] Add fdomain version
This is mostly straightforward. The biggest changes are around encoding as
FDomain packs handles for writing in a slightly different way. Specifically,
regular handles have a HandleDisposition struct which contains a rights field,
an object type, and a HandleOp containing the handle. FDomain moves the rights
into the HandleOp and ditches the HandleDisposition struct.

Change-Id: I6a6a69647b21285fe2b740892231ed96766a612e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1380442
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>