[kernel][phys] Add verbose logging of VM handoff data
When kernel.phys.verbose is set, log full details about all the
VMOs and VMARs in the VM handoff data at the end.  Reuse the same
logging code in the kernel proper when it consumes the data, so
it's easy to compare by eyeball in the logging to see that the VM
handoff data was still intact when the VM system finally used it.

Change-Id: I940b1317dce0712c131cf4777b37a8d9534df0fd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1378621
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>