[shutdown] Fix early stop of critical-services
This change addresses an issue with the critical-services component
which hosts crashsvc and other critical services. This service was
stopping too early in the shutdown of the system, causing exceptions
in late shutdown to not be caught in the crashsvc. This was because
the devfs-with-pkg component is stopped too early. Making this a weak
dependency will break the chain and allow critical-services to live all
the way to the end of shutdown.

Bug: 445524948
Change-Id: I7cd37560d529485ad3a501baf5a4286a7982a9ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1378661
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>