[sestarnix] Apply the memfd-ashmem type override more selectively
We were over-applying the type override to places that did not intend
to create an ashmem-like memfd, leading to attempts to check the
execute permission on the compatibility type. We match a bit more
precisely using the flags passed by Android when creating a memfd
intended to be used as an ashmem replacement.

Bug: 412957798
Change-Id: I1e62ee1d5901e95bf52f5c800fd69b46f9c9b75e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1373925
Commit-Queue: Ambre Williams <ambre@google.com>
Reviewed-by: Wez <wez@google.com>