[starnix][binder] Introduce OperationContext
This patch introduces an `OperationContext` struct to encapsulate the
context of a binder operation. This includes information about the
sending process and thread, the security state, and memory accessors.

This change simplifies the arguments passed to many functions in the
binder driver, making the code cleaner and easier to read.

Change-Id: I639646655b9876d52b70e0069bacaf1cd88edde9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1379761
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Commit-Queue: Benjamin Lerman <qsr@google.com>