[kernel][thread] Enable join_test in k ut intensive_thread
Enable join_test after it was disabled when k thread was migrated to k
ut intensive_thread. It was not originally enabled because it failed.
The failure was due to Detach causing a kernel panic when called on a
dead thread.

This CL fixes the kernel panic by dropping the entire chain lock
transaction before calling Join. Since Join creates its own chain lock
transaction, this change prevents attempting to create a nested
transaction on the thread, which is not allowed.

Bug: 439836795, 447656835
Change-Id: I1cebcdbf2b8ec4a627eef062d4b6ddd42edb4cc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1376501
Commit-Queue: Ethan Williams <ethanws@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>