[ffx] Prevent multiple simultaneous connections
The Mutex in Resolution was incorrectly used -- it was released too
early, so while it protected the field in the Resolution struct, it
didn't prevent connections (and the other lazily-initialized fields)
from being constructed concurrently -- only to be overwritten if another
task constructed the same connection later.

The current version is held over the entire construction of the
connection. Since this includes await points, we need an async Mutex.
We also need to fix the other elements of the Resolution that can
be asynchronously constructed: the rcs_proxy and identify_host_response.

Note that this still does not address all concurrent initialization
issues. We are building multiple Resolutions, overwriting previous
ones in the FhoConnectionBehavior::DirectConnector. This will cause
us to initiate discovery multiple times.

Change-Id: I6a6a696429543399e12565c36929d8281923680d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1379856
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>