[gpt] Ensure metadata has appropriate flushes
The GPT spec requires that secondary and primary metadata are updated in
a particular order.  Enforce this ordering with flushes.

A barrier would also be appropriate, but this is left as a TODO due to
fxbug.dev/416348380.

To test this, I extended VmoBackedServer with write-tracking facilities,
to allow tests to reorder (or otherwise mutate) writes.

Fixed: 446184531
Multiply: gpt-rust-tests
Change-Id: Ic50f63cd33e7f192b38a6f9d80d23c9e6d1ea632
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1370970
Reviewed-by: Nikita Jindal <nikitajindal@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>